Module dcl_crypto::account

Structs

Account

A Struct that allows us to sign messages and serialize and deserialize it easily

Address

EIP1271Signature

And

EphemeralPayload

An EphemeralPayload is a message that delegates the right to sign a message to a specific address until an expiration date.

Expiration

PersonalSignature

Enums

DecodeHexError

An error that can occur when decoding a hexadecimal string

EphemeralPayloadError

Constants

PERSONAL_SIGNATURE_SIZE

Traits

Signer

A trait for signing messages with an associated address.

Functions

decode

Decodes a hex string prefixed with 0x into raw bytes.

decode_to_slice

Decode a hex string prefixed with 0x into a mutable bytes slice.

Type Aliases

EIP1654Signature

Alias of EIP1271Signature See https://eips.ethereum.org/EIPS/eip-1271 See https://github.com/ethereum/EIPs/issues/1654