Struct dcl_crypto::account::Account

pub struct Account(_);

A Struct that allows us to sign messages and serialize and deserialize it easily

Implementations

impl Account

pub fn random() -> Self

Creates a new account generating a random private key.

pub fn from_rng<R: Rng + ?Sized>(r: &mut R) -> Self

Creates a new account using a custom RNG (Random Number Generator) to create the private key.